Make package name lowercase
